+++ src/main/java/sypztep/dominatus/client/payload/AddTextParticlesPayloadS2C.java
@@ -11,27 +11,27 @@ import net.minecraft.server.network.ServerPlayerEntity;
 import sypztep.dominatus.Dominatus;
 import sypztep.dominatus.client.util.TextParticleProvider;
 
-public record AddTextParticlesPayload(int entityId,int selector) implements CustomPayload {
-    public static final Id<AddTextParticlesPayload> ID = new Id<>(Dominatus.id("add_text_particle"));
-    public static final PacketCodec<PacketByteBuf, AddTextParticlesPayload> CODEC = PacketCodec.tuple(
+public record AddTextParticlesPayloadS2C(int entityId, int selector) implements CustomPayload {
+    public static final Id<AddTextParticlesPayloadS2C> ID = new Id<>(Dominatus.id("add_text_particle"));
+    public static final PacketCodec<PacketByteBuf, AddTextParticlesPayloadS2C> CODEC = PacketCodec.tuple(
             PacketCodecs.VAR_INT,
-            AddTextParticlesPayload::entityId,
+            AddTextParticlesPayloadS2C::entityId,
             PacketCodecs.UNSIGNED_SHORT,
-            AddTextParticlesPayload::selector,
-            AddTextParticlesPayload::new
+            AddTextParticlesPayloadS2C::selector,
+            AddTextParticlesPayloadS2C::new
     );
 
     public static void send(ServerPlayerEntity player, int entityId, TextParticleProvider selector) {
-        ServerPlayNetworking.send(player, new AddTextParticlesPayload(entityId, selector.getFlag()));
+        ServerPlayNetworking.send(player, new AddTextParticlesPayloadS2C(entityId, selector.getFlag()));
     }
 
     @Override
     public Id<? extends CustomPayload> getId() {
         return ID;
     }
-    public static class Receiver implements ClientPlayNetworking.PlayPayloadHandler<AddTextParticlesPayload> {
+    public static class Receiver implements ClientPlayNetworking.PlayPayloadHandler<AddTextParticlesPayloadS2C> {
         @Override
-        public void receive(AddTextParticlesPayload payload, ClientPlayNetworking.Context context) {
+        public void receive(AddTextParticlesPayloadS2C payload, ClientPlayNetworking.Context context) {
             Entity entity = context.player().getWorld().getEntityById(payload.entityId());
             if (entity != null) {
                 TextParticleProvider.handleParticle(entity, payload.selector());
